+++ src/main/java/lotto/Application.java
@@ -1,8 +1,11 @@
 package lotto;
 
 public class Application {
-
     public static void main(String[] args) {
-        Controller.startGame();
+        try {
+            LottoGame.startGame();
+        } catch (IllegalArgumentException e) {
+            System.out.println(e.getMessage());
+        }
     }
 }
