+++ src/main/java/lotto/Lotto.java
@@ -1,8 +1,11 @@
 package lotto;
 
+import camp.nextstep.edu.missionutils.Console;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 public class Lotto {
@@ -15,8 +18,44 @@ public class Lotto {
 
     private void validate(List<Integer> numbers) {
         if (numbers.size() != 6) {
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException("[ERROR] 6개의 숫자를 입력해야 합니다.");
         }
+        if (!hasDuplication(numbers)) {
+            throw new IllegalArgumentException("[ERROR] 중복이 존재합니다.");
+        }
+        if (!hasValidScope(numbers)) {
+            throw new IllegalArgumentException("[ERROR] 로또 번호는 1부터 45 사이의 숫자여야 합니다.");
+        }
+    }
+
+    public static Lotto inputWinningNumbers() {
+        System.out.println(Messages.INPUT_WINNING_NUMBERS_MESSAGE);
+        String numbers = Console.readLine();
+        List<Integer> winningNumbers = new ArrayList<>();
+        for (String number : numbers.split(",")) {
+            winningNumbers.add(Integer.parseInt(number));
+        }
+        return new Lotto(winningNumbers);
+    }
+
+    public static void validateBonusNumber(int bonusNumber, Lotto winningNumbers) {
+        if (winningNumbers.numbers.contains(bonusNumber)) {
+            throw new IllegalArgumentException("[ERROR] 중복된 수가 존재합니다.");
+        }
+    }
+
+    public static boolean hasDuplication(List<Integer> numbers) {
+        Set<Integer> elements = new HashSet<>(numbers);
+        return elements.size() == 6;
+    }
+
+    public static boolean hasValidScope(List<Integer> numbers) {
+        for (Integer number : numbers) {
+            if (!(0 < number && number < 46)) {
+                return false;
+            }
+        }
+        return true;
     }
 
     public static List<Lotto> getLottoList(int money) {
@@ -27,17 +66,16 @@ public class Lotto {
         return lottoList;
     }
 
-    public static Lotto generateLottoList() {
+    private static Lotto generateLottoList() {
         List<Integer> numbers = Randoms.pickUniqueNumbersInRange(1, 45, 6);
-        numbers = numbers.stream()
+        return new Lotto(numbers.stream()
                 .sorted()
-                .collect(Collectors.toList());
-        return new Lotto(numbers);
+                .collect(Collectors.toList()));
     }
 
-    public static int compareNumbers(List<Integer> winningNumbers, Lotto lotto) {
+    public static int compareNumbers(Lotto winningNumbers, Lotto lotto) {
         int count = 0;
-        for (Integer number : winningNumbers) {
+        for (Integer number : winningNumbers.numbers) {
             if (lotto.numbers.contains(number)) {
                 count++;
             }
@@ -45,7 +83,7 @@ public class Lotto {
         return count;
     }
 
-    public static boolean isBonusNumber(int bonusNumber, Lotto lotto) {
+    public static boolean hasBonusNumber(int bonusNumber, Lotto lotto) {
         return lotto.numbers.contains(bonusNumber);
     }
