+++ src/main/java/lotto/Messages.java
@@ -3,8 +3,10 @@ package lotto;
 public class Messages {
     public static final String OUTPUT_START_GAME_MESSAGE = "구입금액을 입력해 주세요.";
     public static final String OUTPUT_BUY_MESSAGE = System.lineSeparator() + "%d개를 구매했습니다." +System.lineSeparator();
-    public static final String IMPUT_WINNING_NUMBERS_MESSAGE = System.lineSeparator() + "당첨 번호를 입력해 주세요.";
+    public static final String INPUT_WINNING_NUMBERS_MESSAGE = System.lineSeparator() + "당첨 번호를 입력해 주세요.";
     public static final String INPUT_BONUS_NUMBER_MESSAGE =  System.lineSeparator() +"보너스 번호를 입력해 주세요.";
     public static final String OUTPUT_WINNING_MESSAGE = System.lineSeparator() + "당첨 통계" + System.lineSeparator() + "---";
-    public static final String OUTPUT_RESULT_MESSAGE = System.lineSeparator() + "%d개 일치 (%s원) - %d개" + System.lineSeparator();
+    public static final String OUTPUT_RESULT_MESSAGE = "%d개 일치 (%s원) - %d개" + System.lineSeparator();
+    public static final String OUTPUT_BONUS_MESSAGE = "%d개 일치, 보너스 볼 일치 (%s원) - %d개" + System.lineSeparator();
+    public static final String OUTPUT_YIELD_MESSAGE = "총 수익률은 %.1f%%입니다.";
 }
