+++ src/main/java/lotto/Rank.java
@@ -5,19 +5,18 @@ import java.util.Map;
 import java.util.TreeMap;
 
 public class Rank {
-
     enum Grade {
-        THREE(3, "5,000", false),
-        FOUR(4, "50,000", false),
-        FIVE(5, "1,500,000", false),
-        FIVE_BONUS(5, "30,000,000", true),
-        SIX(6, "2,000,000,000", false);
+        THREE(3, 5000, false),
+        FOUR(4, 50000, false),
+        FIVE(5, 1500000, false),
+        FIVE_BONUS(5, 30000000, true),
+        SIX(6, 2000000000, false);
 
         private final int matchCount;
-        private final String prize;
+        private final long prize;
         private final boolean check;
 
-        Grade(int matchCount, String prize, boolean check) {
+        Grade(int matchCount, long prize, boolean check) {
             this.matchCount = matchCount;
             this.prize = prize;
             this.check = check;
@@ -27,7 +26,7 @@ public class Rank {
             return matchCount;
         }
 
-        public String getPrize() {
+        public long getPrize() {
             return prize;
         }
 
@@ -36,8 +35,16 @@ public class Rank {
         }
     }
 
-    public static Map<Integer, Integer> getResult (List<Integer> winningNumbers, List<Lotto> lottoList, int bonusNumber) {
+    public static Map<Integer, Integer> initializeResult() {
         Map<Integer, Integer> result = new TreeMap<>();
+        for (int key = 3; key < 8; key++) {
+            result.put(key, 0);
+        }
+        return result;
+    }
+
+    public static Map<Integer, Integer> getResult (Lotto winningNumbers, List<Lotto> lottoList, int bonusNumber) {
+        Map<Integer, Integer> result = initializeResult();
         for (Lotto lotto : lottoList) {
             int numberKey = Lotto.compareNumbers(winningNumbers, lotto);
             result.putAll(putNumberKey(numberKey, lotto, bonusNumber));
@@ -47,10 +54,10 @@ public class Rank {
 
     public static Map<Integer, Integer> putNumberKey (int key, Lotto lotto, int bonusNumber) {
         Map<Integer, Integer> result = new TreeMap<>();
-        if (key > 3) {
+        if (key > 2) {
             result.put(key, result.getOrDefault(key, 0) + 1);
         }
-        if (key == 5 && Lotto.isBonusNumber(bonusNumber, lotto)) {
+        if (key == 5 && Lotto.hasBonusNumber(bonusNumber, lotto)) {
             result.put(7, result.getOrDefault(7, 0) + 1);
             result.put(key, result.getOrDefault(key, 0) - 1);
         }
