+++ src/test/java/lotto/LottoGameTest.java
@@ -1,0 +1,15 @@
+package lotto;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class ControllerTest {
+
+    @Test
+    @DisplayName("수익률 출력 확인하기")
+    void getYield() {
+
+    }
+}
