+++ src/test/java/lotto/RankTest.java
@@ -1,4 +1,4 @@
 import static org.junit.jupiter.api.Assertions.*;
-class ControllerTest {
+class RankTest {
   
 }
